git subrepo clone --force git://github.com/perlpilot/p6-File-Temp.git perlpilot/p6-File-Temp
subrepo:
  subdir:   "perlpilot/p6-File-Temp"
  merged:   "146b2ec"
upstream:
  origin:   "git://github.com/perlpilot/p6-File-Temp.git"
  branch:   "master"
  commit:   "146b2ec"
git-subrepo:
  version:  "0.2.1"
  origin:   "git@github.com:ingydotnet/git-subrepo.git"
  commit:   "b136967"

+++ perlpilot/p6-File-Temp/lib/File/Temp.pm
@@ -1,5 +1,7 @@
 module File::Temp:ver<0.02>;
 
+use File::Directory::Tree;
+
 # Characters used to create temporary file/directory names
 my @filechars = 'a'..'z', 'A'..'Z', 0..9, '_';
 constant MAX-RETRIES = 10;
@@ -32,9 +34,39 @@ sub tempfile (
     return ();
 }
 
+our sub tempdir (
+    $tmpl? = '*' x 10,          # positional template
+    :$tempdir? = $*TMPDIR,      # where to create tempdir
+    :$prefix? = '',             # dirname prefix
+    :$unlink?  = 1,             # remove when program exits?
+    :$template = $tmpl          # required named template
+) is export {
+
+    my $count = MAX-RETRIES;
+    while ($count--) {
+        my $tempdir_name = $template;
+        $tempdir_name ~~ s/ '*' ** 4..* /{ gen-random($/.chars) }/;
+        my $dirpath = "$tempdir/$prefix$tempdir_name";
+        next if $dirpath.IO ~~ :e;
+        try { CATCH { next }; mkdir($dirpath) };
+        push @open-files, $dirpath if $unlink;
+        return $dirpath;
+    }
+    return ();
+}
+
 END {
     for @open-files -> $f {
-        unlink($f);
+        next unless $f.IO ~~ :e; # maybe warn here
+
+        if $f.IO ~~ :f
+        {
+            unlink($f);
+        }
+        elsif $f.IO ~~ :d
+        {
+            rmtree($f);
+        }
     }
 }
